change input coord type to float from int

+++ src/main.cpp
@@ -55,10 +55,10 @@ int main(int argc, const char **argv)
     // TODO 1: Declare floats `start_x`, `start_y`, `end_x`, and `end_y` and get
     // user input for these values using std::cin. Pass the user input to the
     // RoutePlanner object below in place of 10, 10, 90, 90.
-    int start_x;
-    int start_y;
-    int end_x;
-    int end_y;
+    float start_x;
+    float start_y;
+    float end_x;
+    float end_y;
     std::cout << "Input start position x, followed by y:\n";
     std::cin >> start_x;
     std::cin >> start_y;
